Add files via upload
A Solar grass cutter is a machine that used spiral BLDC to cut a lawn and as well as frame at 
a faster rate. Solar grass cutter as operate automatic and motor driven. Reel cutter of the grass 
cutter is given adjust height and distance. Power consumption becomes fundamental need for 
future. Solar grass cutter is a very useful device which is very easy and simple in 
construction. It is used to maintain and upkeep lawns in gardens, schools, college’s, frames 
etc. We have made some changes in the existing machine to make its application easy and 
reliable at reduced cost. Our main aim to project on this topic is to pollution control and 
unskilled person can also operate easily and maintain the lawn very fine and uniform surface 
look. In our project, solar grass cutter is used to cut the different grasses for the different 
application and energy conversions